refactor: Restructure map-related Sidekiq workers and extract their core logic into dedicated classes.
