+++ lib/ruby_reactor/sidekiq_workers/map_collector_worker.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RubyReactor
+  module SidekiqWorkers
+    class MapCollectorWorker
+      include ::Sidekiq::Worker
+
+      def perform(arguments)
+        RubyReactor::Map::Collector.perform(arguments)
+      end
+    end
+  end
+end
