+++ lib/ruby_reactor/sidekiq_workers/map_element_worker.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RubyReactor
+  module SidekiqWorkers
+    class MapElementWorker
+      include ::Sidekiq::Worker
+
+      def perform(arguments)
+        RubyReactor::Map::ElementExecutor.perform(arguments)
+      end
+    end
+  end
+end
