+++ lib/ruby_reactor/sidekiq_workers/map_execution_worker.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require "sidekiq"
+
+module RubyReactor
+  module SidekiqWorkers
+    class MapExecutionWorker
+      include ::Sidekiq::Worker
+
+      def perform(arguments)
+        RubyReactor::Map::Execution.perform(arguments)
+      end
+    end
+  end
+end
